data-analyzer-llm/requirements.txt
fastapi
uvicorn
python-docx
python-pptx
beautifulsoup4
markdownify
PyMuPDF
